Refactoring loading of the LDAP object
