+++ src/Identifier/Backend/Ldap.php
@@ -10,10 +10,28 @@
  * @link          http://cakephp.org CakePHP(tm) Project
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Authentication\Identifier;
+namespace Authentication\Identifier\Backend;
 
-interface LdapInterface
+use RuntimeException;
+
+/**
+ * Provides a very thin OOP wrapper around the ldap_* functions.
+ *
+ * We don't need and want a huge LDAP lib for our purpose.
+ *
+ * But this makes it easier to unit test code that is using ldap because we can
+ * mock it very easy. It also provides some convenience.
+ */
+class Ldap implements LdapInterface
 {
+
+    /**
+     * LDAP Object
+     *
+     * @var object
+     */
+    protected $_connection;
+
     /**
      * Bind to LDAP directory
      *
@@ -21,7 +39,10 @@ interface LdapInterface
      * @param string $password Bind password
      * @return bool
      */
-    public function bind($bind, $password);
+    public function bind($bind, $password)
+    {
+        return ldap_bind($this->getConnection(), $bind, $password);
+    }
 
     /**
      * Get the LDAP connection
@@ -29,7 +50,14 @@ interface LdapInterface
      * @return mixed
      * @throws \RuntimeException If the connection is empty
      */
-    public function getConnection();
+    public function getConnection()
+    {
+        if (empty($this->_connection)) {
+            throw new RuntimeException('You are not connected to a LDAP server.');
+        }
+
+        return $this->_connection;
+    }
 
     /**
      * Connect to an LDAP server
@@ -38,7 +66,10 @@ interface LdapInterface
      * @param int $port Port
      * @return void
      */
-    public function connect($host, $port);
+    public function connect($host, $port)
+    {
+        $this->_connection = ldap_connect($host, $port);
+    }
 
     /**
      *  Set the value of the given option
@@ -47,7 +78,10 @@ interface LdapInterface
      * @param mixed $value The new value for the specified option
      * @return void
      */
-    public function setOption($option, $value);
+    public function setOption($option, $value)
+    {
+        ldap_set_option($this->getConnection(), $option, $value);
+    }
 
     /**
      * Get the current value for given option
@@ -55,12 +89,21 @@ interface LdapInterface
      * @param int $option Option to get
      * @return mixed This will be set to the option value.
      */
-    public function getOption($option);
+    public function getOption($option)
+    {
+        ldap_get_option($this->getConnection(), $option, $returnValue);
+
+        return $returnValue;
+    }
 
     /**
      * Unbind from LDAP directory
      *
      * @return void
      */
-    public function unbind();
+    public function unbind()
+    {
+        ldap_unbind($this->_connection);
+        $this->_connection = null;
+    }
 }
